feat: add sidebar, toast viewport, and world timer components
- Implemented Sidebar component to display agents grouped by project with selection functionality.
- Created ToastViewport component for displaying toast notifications with dismiss functionality.
- Added WorldTimer component to show elapsed time since a given start time.

feat: introduce hooks for dashboard details and session management

- Added useDashboardDetails hook to fetch and manage dashboard details for agents.
- Implemented useSessionDetail hook for fetching session details with polling capability.
- Created useStableProjectColors hook to maintain consistent colors for project groups.

feat: enhance dashboard view model with utility functions

- Introduced various utility functions for managing project and tool data, including grouping agents, formatting project names, and handling provider icons.
- Added tests for dashboard view model functions to ensure correctness.

feat: implement text size presets for consistent UI styling

- Created text size presets and utility functions for mapping text scale values to preset keys.
- Added tests to validate text size preset functionality.

test: add tests for collector publisher and snapshot functionality

- Implemented tests for collector publisher to ensure snapshot publishing behavior and fingerprint consistency.
- Added tests for snapshot normalization and building collector snapshots.

feat: develop hubreceiver routes and websocket management

- Implemented hubreceiver routes for handling snapshot submissions and session details.
- Created websocket manager for broadcasting updates to connected clients.
- Added tests for hubreceiver routes and websocket functionality to ensure reliability.
