+++ claudeville/src/presentation/App.ts
@@ -13,6 +13,7 @@ import { SessionWatcher } from '../application/SessionWatcher.js';
 import { NotificationService } from '../application/NotificationService.js';
 import { getNameMode, setNameMode } from '../config/agentNames.js';
 import { getBubbleConfig, updateBubbleConfig } from '../config/bubbleConfig.js';
+import { TEXT_SIZE_PRESETS, getTextSizePresetKey } from './shared/textSizePresets.js';
 
 import { TopBar } from './shared/TopBar.js';
 import { Sidebar } from './shared/Sidebar.js';
@@ -198,25 +199,12 @@ export class App {
         const btn = document.getElementById('btnSettings');
         if (!btn) return;
 
-        /**
-         * TEXT_SIZE_PRESETS — drives both:
-         *   - --text-scale on :root (scales ALL UI elements)
-         *   - statusFontSize / chatFontSize in bubbleConfig (scales world-mode bubbles)
-         */
-        const TEXT_SIZE_PRESETS = [
-            { key: 'small',  labelKey: 'bubbleSmall',       textScale: 0.8, statusFontSize: 10, maxWidth: 160, bubbleH: 22, paddingH: 18 },
-            { key: 'medium', labelKey: 'bubbleMedium',      textScale: 1.0, statusFontSize: 14, maxWidth: 260, bubbleH: 28, paddingH: 24 },
-            { key: 'large',  labelKey: 'bubbleLarge',       textScale: 1.25, statusFontSize: 20, maxWidth: 360, bubbleH: 38, paddingH: 32 },
-            { key: 'xlarge', labelKey: 'bubbleExtraLarge',   textScale: 1.5, statusFontSize: 28, maxWidth: 480, bubbleH: 52, paddingH: 44 },
-        ];
-
         const sizeLabel = (preset: { key: string; labelKey: string }) =>
             i18n.t(preset.labelKey || `bubble${preset.key.charAt(0).toUpperCase() + preset.key.slice(1)}`);
 
         const buildForm = (currentMode: string) => {
             const cfg = getBubbleConfig();
-            const currentScaleKey = TEXT_SIZE_PRESETS.find(p => p.textScale === cfg.textScale)?.key
-                || (cfg.textScale < 0.9 ? 'small' : cfg.textScale < 1.1 ? 'medium' : cfg.textScale < 1.4 ? 'large' : 'xlarge');
+            const currentScaleKey = getTextSizePresetKey(cfg.textScale);
 
             const sizeBtns = TEXT_SIZE_PRESETS.map(p => {
                 const active = currentScaleKey === p.key ? ' settings-lang-btn--active' : '';
