+++ hubreceiver/server.ts
@@ -2,174 +2,28 @@ require('../load-local-env.ts');
 
 const http = require('http');
 const { applySnapshot, getCurrentState, getSessionDetail, getHistory, defaultUsage } = require('./state');
+const { createHubreceiverRequestHandler } = require('./routes');
+const { createHubWebSocketManager } = require('./ws');
 
 const PORT = Number(process.env.HUB_PORT || 3030);
 const AUTH_TOKEN = process.env.HUB_AUTH_TOKEN || 'dev-secret';
 const MAX_SNAPSHOT_BYTES = Number(process.env.MAX_SNAPSHOT_BYTES || 10 * 1024 * 1024); // 10 MB default
 
-const wsClients = new Set();
-
-const { setCorsHeaders, sendJson, sendError, safeLimit, readBoundedBody } = require('../shared/http-utils');
-const { createWebSocketFrame, computeAcceptKey } = require('../shared/ws-utils');
-const { wsSend, wsBroadcast } = require('../shared/ws-helpers');
-
-const { MIME_TYPES } = require('../shared/mime-types');
-
-function handleWebSocketUpgrade(req, socket) {
-  const key = req.headers['sec-websocket-key'];
-  if (!key) {
-    socket.destroy();
-    return;
-  }
-
-  const acceptKey = computeAcceptKey(key);
-  const responseStr =
-    'HTTP/1.1 101 Switching Protocols\r\n' +
-    'Upgrade: websocket\r\n' +
-    'Connection: Upgrade\r\n' +
-    'Sec-WebSocket-Accept: ' + acceptKey + '\r\n' +
-    '\r\n';
-
-  socket.write(responseStr, () => {
-    wsClients.add(socket);
-    wsSend(socket, buildWsPayload('init'));
-  });
-
-  socket.on('close', () => {
-    wsClients.delete(socket);
-  });
-
-  socket.on('error', () => {
-    wsClients.delete(socket);
-  });
-}
-
-function buildWsPayload(type) {
-  const state = getCurrentState();
-  return {
-    type,
-    sessions: state.sessions,
-    teams: state.teams,
-    taskGroups: state.taskGroups,
-    providers: state.providers,
-    usage: state.usage,
-    timestamp: state.timestamp || Date.now(),
-  };
-}
-
-function maybeGetAuthToken(req) {
-  const header = req.headers.authorization || '';
-  return header.replace(/^Bearer /i, '');
-}
-
-const server = http.createServer((req, res) => {
-  if (req.method === 'OPTIONS') {
-    setCorsHeaders(res);
-    res.writeHead(204);
-    res.end();
-    return;
-  }
-
-  const url = new URL(req.url, `http://${req.headers.host}`);
-  const pathname = url.pathname;
-
-  if (req.method === 'GET' && pathname === '/health') {
-    sendJson(res, 200, { ok: true, collectors: getCurrentState().sessions.length });
-    return;
-  }
-
-  if (req.method === 'POST' && pathname === '/api/collector/snapshot') {
-    if (maybeGetAuthToken(req) !== AUTH_TOKEN) {
-      sendError(res, 401, 'unauthorized');
-      return;
-    }
-
-    readBoundedBody(req, MAX_SNAPSHOT_BYTES)
-      .then(({ body }) => {
-        try {
-          const snapshot = JSON.parse(body || '{}');
-          const state = applySnapshot(snapshot);
-          wsBroadcast({
-            ...buildWsPayload('update'),
-            sessions: state.sessions,
-            teams: state.teams,
-            taskGroups: state.taskGroups,
-            providers: state.providers,
-            usage: state.usage,
-            timestamp: state.timestamp || Date.now(),
-          }, wsClients);
-          console.log(`[hubreceiver] snapshot accepted ${Buffer.byteLength(body, 'utf8')} bytes → ${state.sessions.length} sessions`);
-          sendJson(res, 200, { ok: true, sessions: state.sessions.length });
-        } catch (error) {
-          console.error(`[hubreceiver] snapshot parse error (${Buffer.byteLength(body, 'utf8')} bytes): ${error instanceof Error ? error.message : error}`);
-          sendError(res, 400, error instanceof Error ? error.message : 'invalid snapshot');
-        }
-      })
-      .catch((err) => {
-        if (err && err.statusCode === 413) {
-          console.error(`[hubreceiver] snapshot rejected — ${err.message}`);
-          sendError(res, 413, err.message);
-        } else {
-          console.error(`[hubreceiver] snapshot read error: ${err}`);
-          sendError(res, 400, 'failed to read snapshot body');
-        }
-      });
-    return;
-  }
-
-  if (req.method === 'GET' && pathname === '/api/sessions') {
-    const state = getCurrentState();
-    sendJson(res, 200, { sessions: state.sessions, count: state.sessions.length, timestamp: state.timestamp });
-    return;
-  }
-
-  if (req.method === 'GET' && pathname === '/api/session-detail') {
-    const sessionId = url.searchParams.get('sessionId');
-    const provider = url.searchParams.get('provider') || 'claude';
-    if (!sessionId) {
-      sendError(res, 400, 'sessionId 필수');
-      return;
-    }
-    sendJson(res, 200, getSessionDetail(sessionId, provider));
-    return;
-  }
-
-  if (req.method === 'GET' && pathname === '/api/teams') {
-    const state = getCurrentState();
-    sendJson(res, 200, { teams: state.teams, count: state.teams.length });
-    return;
-  }
-
-  if (req.method === 'GET' && pathname === '/api/tasks') {
-    const state = getCurrentState();
-    sendJson(res, 200, { taskGroups: state.taskGroups, totalGroups: state.taskGroups.length });
-    return;
-  }
-
-  if (req.method === 'GET' && pathname === '/api/providers') {
-    const state = getCurrentState();
-    sendJson(res, 200, { providers: state.providers, count: state.providers.length });
-    return;
-  }
-
-  if (req.method === 'GET' && pathname === '/api/usage') {
-    const state = getCurrentState();
-    sendJson(res, 200, state.usage || defaultUsage());
-    return;
-  }
-
-  if (req.method === 'GET' && pathname === '/api/history') {
-    const limit = safeLimit(url.searchParams.get('lines'));
-    sendJson(res, 200, { entries: getHistory(limit) });
-    return;
-  }
-
-  sendError(res, 404, 'Not Found');
-});
+const wsManager = createHubWebSocketManager(getCurrentState);
+const server = http.createServer(createHubreceiverRequestHandler({
+  applySnapshot,
+  getCurrentState,
+  getSessionDetail,
+  getHistory,
+  defaultUsage,
+  wsManager,
+  authToken: AUTH_TOKEN,
+  maxSnapshotBytes: MAX_SNAPSHOT_BYTES,
+}));
 
 server.on('upgrade', (req, socket) => {
   if (req.headers.upgrade && req.headers.upgrade.toLowerCase() === 'websocket') {
-    handleWebSocketUpgrade(req, socket);
+    wsManager.handleUpgrade(req, socket);
     return;
   }
   socket.destroy();
